fixes page bugs (calling non-existant functions) and seperates the 24hrs and facetHistory demo pages and functions
