+++ views/index.hbs
@@ -10,6 +10,9 @@
 			<li>
 				<a href="24hrs">24hrs of news</a>
 			</li>
+			<li>
+				<a href="facetHistory">facetHistory</a>
+			</li>
 		</ul>
 	</body>
 </html>
