+++ views/twentyfourhrs.hbs
@@ -26,27 +26,6 @@
 		  GET /topics/last7days, returns JSON, e.g. <a href="/24hrs/topics/last7days">/topics/last7days</a>
 		</p>
 
-		<p>
-		  GET /facetHistory/:facet/:period/:interval/:numInterval/:maxFacets, returns JSON, e.g. <a href="/24hrs/facetHistory/topics/days/1/5/10">/facetHistory/topics/days/1/5/10</a>
-		</p>
-		<ul>
-			<li>
-				<b>facet</b> : string : facet type to query [topic, organisation, people] matching id facet auto included e.g. people &amp; peopleId 
-			</li>
-			<li>
-				<b>period</b> : string : time period to query [minutes, hours, days]
-			</li>
-			<li>
-				<b>interval</b> : integer : quantity of the time period to query [1, 10]
-			</li>
-			<li>
-				<b>numInterval</b> : integer : number of intervals [1, 10]
-			</li>
-			<li>
-				<b>maxFacets</b> : integer : number of facets to retrun results for [0, 100]
-			</li>
-		</ul>
-
 		<h2>Pages</h2>
 
 		<ul>
@@ -55,9 +34,6 @@
 			</li>
 			<li>
 				<a href="/24hrs/2">24hrs v2 - Image display</a>
-			</li>
-			<li>
-				<a href="/24hrs/3">24hrs v3 - </a>
 			</li>
 		</ul>
 	</body>
