--- views/twentyfourhrs/three.hbs
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-	{{#> head title='24hrs v3 - ' subtitle='' }}
-		<style type="text/css">
-		</style>
-	{{/head}}
-	<body>
-		{{> header title='24hrs v3 - ' subtitle=''}}
-		
-		<main>
-			<!-- results to go here -->
-		</main>
-
-		<script type="text/javascript" src="/static/js/main.js"></script>
-		<script>
-			fetch('/24hrs/topics/last7days')
-				.then(status)
-				.then(json)
-				.then(function(data) {
-					console.log('Request succeeded with JSON response', data);
-					displayInDOM(data);
-				}).catch(function(error) {
-					console.log('Request failed', error);
-			});
-
-			function displayInDOM(result){
-				if(result != null)
-				{
-					const main = document.getElementsByTagName('main')[0];
-					main.innerHTML = JSON.stringify(result);
-				}
-				else {
-					console.log("No usable JSON returned");
-				}
-			}
-		</script>
-	</body>
-</html>
